balltracker_trainer/ball-detector/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ball-detector"
version = "0.1.0"
description = "Fine-tune YOLOv11 for cricket ball detection with SAHI and ByteTrack"
requires-python = ">=3.10"
dependencies = [
    "ultralytics>=8.3",
    "opencv-python>=4.8",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
    "tqdm>=4.65",
    "sahi>=0.11",
    "lap>=0.4",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
]
